+++ packages/demo/index.js
@@ -1,6 +1,9 @@
 module.exports = {
   barrage_hell: require('./barrages/barrage_hell'),
+  deep_vortex: require('./barrages/deep_vortex'),
   dream_express: require('./barrages/dream_express'),
+  // peta_flare: require('./barrages/peta_flare'),
   purification_of_filth: require('./barrages/purification_of_filth'),
+  rainy_night: require('./barrages/rainy_night'),
   wave_and_particle: require('./barrages/wave_and_particle'),
 }