+++ packages/touhou/src/index.ts
@@ -1,11 +1,13 @@
 import templates from './templates'
 import bullets from './bullets'
+import th11 from './touhou/th11'
 import th12 from './touhou/th12'
 import th15 from './touhou/th15'
 
 export {
   templates,
   bullets,
+  th11,
   th12,
   th15,
 }